split into old test suites
