+++ core/controllers/old_suggestion_test.py
@@ -328,120 +328,3 @@ class SuggestionUnitTests(test_utils.GenericTestBase):
                     suggestion_models.TARGET_TYPE_EXPLORATION, self.EXP_ID)
                 )['suggestions']
             self.assertEqual(len(suggestions), 2)
-
-
-class QuestionSuggestionTests(test_utils.GenericTestBase):
-
-    AUTHOR_EMAIL = 'author@example.com'
-    AUTHOR_EMAIL_2 = 'author2@example.com'
-
-    # Needs to be 12 characters long.
-    SKILL_ID = 'skill1234567'
-
-    SKILL_DESCRIPTION = 'skill to link question to'
-    TOPIC_ID = 'topic'
-
-    def setUp(self):
-        super(QuestionSuggestionTests, self).setUp()
-        self.signup(self.ADMIN_EMAIL, self.ADMIN_USERNAME)
-        self.signup(self.AUTHOR_EMAIL, 'author')
-        self.admin_id = self.get_user_id_from_email(self.ADMIN_EMAIL)
-        self.author_id = self.get_user_id_from_email(self.AUTHOR_EMAIL)
-        self.set_admins([self.ADMIN_USERNAME])
-        self.save_new_skill(
-            self.SKILL_ID, self.admin_id, self.SKILL_DESCRIPTION)
-        self.question_dict = {
-            'question_state_data': self._create_valid_question_data(
-                'default_state').to_dict(),
-            'language_code': 'en',
-            'question_state_schema_version': (
-                feconf.CURRENT_STATES_SCHEMA_VERSION)
-        }
-        with self.swap(constants, 'USE_NEW_SUGGESTION_FRAMEWORK', True):
-            with self.swap(
-                constants, 'ENABLE_GENERALIZED_FEEDBACK_THREADS', False):
-                self.login(self.AUTHOR_EMAIL)
-                response = self.testapp.get(feconf.CREATOR_DASHBOARD_URL)
-                csrf_token = self.get_csrf_token_from_response(response)
-
-                self.post_json(
-                    '%s/' % feconf.GENERAL_SUGGESTION_URL_PREFIX, {
-                        'suggestion_type': (
-                            suggestion_models.SUGGESTION_TYPE_ADD_QUESTION),
-                        'target_type': suggestion_models.TARGET_TYPE_TOPIC,
-                        'target_id': self.TOPIC_ID,
-                        'target_version_at_submission': 1,
-                        'change': {
-                            'cmd': (
-                                question_domain
-                                .CMD_CREATE_NEW_FULLY_SPECIFIED_QUESTION),
-                            'question_dict': self.question_dict,
-                            'skill_id': self.SKILL_ID
-                        },
-                        'description': 'Add new question to skill'
-                    }, csrf_token=csrf_token)
-                self.logout()
-
-    def test_query_question_suggestions(self):
-        with self.swap(constants, 'USE_NEW_SUGGESTION_FRAMEWORK', True):
-            suggestions = self.get_json(
-                '%s?suggestion_type=%s' % (
-                    feconf.GENERAL_SUGGESTION_LIST_URL_PREFIX,
-                    suggestion_models.SUGGESTION_TYPE_ADD_QUESTION)
-                )['suggestions']
-            self.assertEqual(len(suggestions), 1)
-            suggestion = suggestions[0]
-            self.assertEqual(
-                suggestion['suggestion_type'],
-                suggestion_models.SUGGESTION_TYPE_ADD_QUESTION)
-            self.assertEqual(suggestion['target_id'], self.TOPIC_ID)
-            self.assertEqual(
-                suggestion['target_type'], suggestion_models.TARGET_TYPE_TOPIC)
-            self.assertEqual(
-                suggestion['change']['cmd'],
-                question_domain.CMD_CREATE_NEW_FULLY_SPECIFIED_QUESTION)
-
-    def test_accept_question_suggestion(self):
-        with self.swap(constants, 'USE_NEW_SUGGESTION_FRAMEWORK', True):
-            with self.swap(
-                constants, 'ENABLE_GENERALIZED_FEEDBACK_THREADS', False):
-                suggestion_to_accept = self.get_json(
-                    '%s?suggestion_type=%s' % (
-                        feconf.GENERAL_SUGGESTION_LIST_URL_PREFIX,
-                        suggestion_models.SUGGESTION_TYPE_ADD_QUESTION)
-                    )['suggestions'][0]
-
-                self.login(self.ADMIN_EMAIL)
-                response = self.testapp.get(feconf.CREATOR_DASHBOARD_URL)
-                csrf_token = self.get_csrf_token_from_response(response)
-                with self.swap(feconf, 'ENABLE_NEW_STRUCTURES', True):
-                    self.put_json('%s/topic/%s/%s' % (
-                        feconf.GENERAL_SUGGESTION_ACTION_URL_PREFIX,
-                        suggestion_to_accept['target_id'],
-                        suggestion_to_accept['suggestion_id']), {
-                            'action': u'accept',
-                            'commit_message': u'commit message',
-                            'review_message': u'This looks good!'
-                        }, csrf_token=csrf_token)
-
-                suggestion_post_accept = self.get_json(
-                    '%s?suggestion_type=%s' % (
-                        feconf.GENERAL_SUGGESTION_LIST_URL_PREFIX,
-                        suggestion_models.SUGGESTION_TYPE_ADD_QUESTION)
-                    )['suggestions'][0]
-                self.assertEqual(
-                    suggestion_post_accept['status'],
-                    suggestion_models.STATUS_ACCEPTED)
-                questions, _ = (
-                    question_services.get_question_summaries_linked_to_skills(
-                        1, [self.SKILL_ID], ''))
-                self.assertEqual(len(questions), 1)
-                self.assertEqual(questions[0].creator_id, self.author_id)
-                self.assertEqual(
-                    questions[0].question_content,
-                    self.question_dict['question_state_data']['content']['html']
-                )
-                thread_messages = feedback_services.get_messages(
-                    suggestion_to_accept['suggestion_id'])
-            last_message = thread_messages[len(thread_messages) - 1]
-            self.assertEqual(last_message.text, 'This looks good!')
